Fix touch and quoting on windows
The tests for this module fail on Windows. Fix two separate but
related problems here. 

First, Windows has no `touch` command, so use `File::Touch::touch()`
instead.

Second, on Windows `'` is not a quoting character in the CMD, so
changed the `system` arguments to use `"` instead.

Tests still don't pass on Windows due to some encoding problems with
Miyagawa/Salmon in that author name on line 29,
but this specific portion now works correctly on Windows.