StopWords is a Pod::Weaver **plugin**, so use -StopWords as the key
For now, remember that StopWords is a Pod::Weaver plugin, and as such, needs a
different prefix than a section.  (This wasn't in my clue list for... ever.)

We'll use the correct key ("-StopWords") and continue to add contributor names
as stopwords until such time as we do not provide any data to Pod::Weaver from
this plugin.  As we're providing contributors to one Pod::Weaver section, we
should also pass this information on to -StopWords, taking care of everything
required rather than passing responsibility elsewhere.